Move interfaces into ts/interfaces
