+++ lib/next-auth.d.ts
@@ -1,5 +1,6 @@
 import NextAuth from "next-auth";
-import { Identification } from "./interfaces";
+
+import { Identification } from "./ts/interfaces/identification.interfaces";
 
 declare module "next-auth" {
   interface Session {
